+++ src/commands/navigate-route-command.js
@@ -1,0 +1,15 @@
+const NavigateTo = require('./navigate-to');
+
+class NavigateRouteCommand extends NavigateTo {
+  execute() {
+    this.namespaces = new Map([
+      ['controllers', ['routes']],
+      ['templates', ['routes']],
+      ['routes', ['controllers', 'router']],
+    ]);
+
+    super.execute();
+  }
+}
+
+module.exports = NavigateRouteCommand;