+++ src/commands/navigate-template-command.js
@@ -1,0 +1,16 @@
+const NavigateTo = require('./navigate-to');
+
+class NavigateTemplateCommand extends NavigateTo {
+  execute() {
+    this.namespaces = new Map([
+      ['controllers', ['templates']],
+      ['components', ['templates']],
+      ['templates', ['controllers', 'components']],
+      ['routes', ['templates']]
+    ]);
+
+    super.execute();
+  }
+}
+
+module.exports = NavigateTemplateCommand;