+++ src/extension.js
@@ -1,6 +1,8 @@
 const vscode = require('vscode');
 const QuickPickBuilder = require('./quick-pick-builder');
-const ToggleJavaScriptHandlebarsCommand = require('./commands/toggle-js-hbs-command');
+const NavigateToTemplateCommand = require('./commands/navigate-template-command');
+const NavigateToRouteCommand = require('./commands/navigate-route-command');
+const NavigateToTestCommand = require('./commands/navigate-test-command');
 const NavigateToActiveComponentCommand = require('./commands/navigate-active-component-command');
 
 function activate(context) {
@@ -20,8 +22,16 @@ function activate(context) {
 
   context.subscriptions.push(disposable);
 
-  context.subscriptions.push(vscode.commands.registerCommand('emberNavigator.toggleJavaScriptHandlebars', function () {
-    new ToggleJavaScriptHandlebarsCommand(vscode.window.activeTextEditor.document.fileName).execute();
+  context.subscriptions.push(vscode.commands.registerCommand('emberNavigator.navigateTemplate', function () {
+    new NavigateToTemplateCommand(vscode.window.activeTextEditor.document.fileName).execute();
+  }));
+
+  context.subscriptions.push(vscode.commands.registerCommand('emberNavigator.navigateRoute', function () {
+    new NavigateToRouteCommand(vscode.window.activeTextEditor.document.fileName).execute();
+  }));
+
+  context.subscriptions.push(vscode.commands.registerCommand('emberNavigator.navigateTest', function () {
+    new NavigateToTestCommand(vscode.window.activeTextEditor.document.fileName).execute();
   }));
 
   context.subscriptions.push(vscode.commands.registerCommand('emberNavigator.navigateToActiveComponent', function () {