+++ src/strategies/implementations/template-strategy.js
@@ -14,6 +14,11 @@ class TemplateNavigationStrategy extends NavigationStrategy {
         path: path.join(this.location.project, this.location.parent, 'controllers', this.location.optDirectories, `${this.location.file}.js`),
         icon: 'javascript',
         namespace: 'controllers'
+      },
+      {
+        path: path.join(this.location.project, this.location.parent, 'routes', this.location.optDirectories, `${this.location.file}.js`),
+        icon: 'javascript',
+        namespace: 'routes'
       }
     ]);